Extracted public address generation into its own module.
