+++ public_key_gen.rb
